feat: Added View for Contacted, Uncontacted, and None
These views were added through a enum FilterType that sends a permeter of each of these views. This allows the ProspectView to be used in the Tabbar with three different context. To start, each of these Tabs represented with a Filter with a different title through the different case items brought in with the Filter switch statement