feat: Added Text field input that generates a QRCode
A text field binded to two State properties that register the user's name and email address were added. The bottom item of the VStack holds a UIimage that is generate from the generateQRCode method written that takes in a String and outputs a UIImage using the imported  CoreImage.CIFilterBuiltins. The inputmessage sets the value and filtered to an outputImage that returns a UIImage. This gives a unique QRCode for each user that puts in there name and email address