feat: Added sheet to scan QRCode
The action added provides a sheet that asks the user for permission to scan the QRCode. When the sheet is run in the simulator the devloper can choose to have a sample QRCode run that will add a sample name and email address to ProspectsView in the first tab. This was implemented through importing CodeScanner spm that is called upon launch of the sheet. The user can scan a QRCode that is registered and have the name and contact information kept in their ProspectsView