feat: Added Save and Add Method
Added a method that calls a save to user defaults. In order for the information to be saved without using a Stringly type, a varible was added to be able to call the save and invalidates the Save method without using the created Add method that takes in the created saveKey varible.